refactor: extract components from app component
