+++ samples/src/moderated-chat/moderated-chat.scss
@@ -60,98 +60,6 @@
   }
 }
 
-.modal {
-  opacity: 0;
-  transition: 0.3s ease-in-out;
-  width: 100%;
-  height: 100%;
-  background-color: var(--app-modal-background);
-  position: fixed;
-  z-index: 100;
-  display: none;
-
-  &.shown {
-    opacity: 1;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-  }
-
-  &_content-area {
-    background-color: var(--app-background);
-    color: var(--app-color);
-    width: 500px;
-    max-width: 100%;
-    border-radius: 25px;
-    padding: 24px;
-  }
-
-  &_actions {
-    text-align: right;
-
-    button {
-      color: var(--app-color);
-      border: none;
-      border-radius: 25px;
-      padding: 10px 20px;
-      background-color: transparent;
-      cursor: pointer;
-
-      &:hover {
-        text-decoration: underline;
-      }
-
-      &.cta {
-        color: var(--app-modal__action__color);
-        background-color: var(--app-modal__action__background);
-      }
-    }
-  }
-}
-
-.file-picker {
-  input[type="file"] {
-    display: none;
-  }
-
-  label {
-    padding-bottom: 5px;
-    vertical-align: middle;
-    cursor: pointer;
-  }
-}
-
-.file-preview {
-  .filename {
-    font-size: 12px;
-  }
-
-  .description {
-    font-weight: bold;
-    font-size: 15px;
-    font-style: normal;
-    margin-top: 42px;
-
-    span {
-      font-weight: normal;
-    }
-  }
-
-  img {
-    height: 120px;
-    border-radius: 10px;
-  }
-
-  .pn-msg-input {
-    --text-field__send__borderRadius: 25px;
-    --text-field__send__padding: 10px 20px;
-    --text-field__textarea__borderRadius: 25px;
-    --text-field__textarea__padding: 12px 24px;
-    padding: 12px 0px 24px 0px;
-    background-color: transparent;
-  }
-}
-
 .app-simple {
   background: var(--app-background);
   color: var(--app-color);
